Fix Forward() livelock once command_id exceeds 65535 (#530)

+++ tx_service/include/remote/remote_cc_request.h
@@ -355,7 +355,7 @@ public:
         }
     }
 
-    uint16_t CommandId()
+    uint32_t CommandId()
     {
         return input_msg_->command_id();
     }
